changes needed to use usvg-0.34 and fontdb-0.14.

+++ src/main.rs
@@ -15,7 +15,7 @@ struct Args {
     output: Option<PathBuf>,
     /// The number of SVG pixels per PDF points.
     #[clap(long, default_value = "72.0")]
-    dpi: f64,
+    dpi: f32,
 }
 
 fn main() {
